1. 调整metric注册逻辑，有等初始化的item，则快速进入一下个循环，不sleep. 2. 通过macro整合不同type的定义； 3. 支持ratio
