+++ stream/src/metric.rs
@@ -46,6 +46,7 @@ macro_rules! define_metrics {
 }
 
 define_metrics!(qps: tx-tx, rx-rx, err-err, cps-cps, kps-kps, conn-conn,noresponse-noresponse, key-key, hit-hit;
-count:conn_num-conn;
-rtt:avg-avg
+num:conn_num-conn;
+rtt:avg-avg;
+ratio:cache-hit
 );
